# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AKIM AI — Аким на 5 часов",
    "description": "Учебная симуляция на синтетических данных. Детерминированные расчёты; AI объясняет готовые факты.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/ai/analyze": {
      "post": {
        "tags": [
          "AI integration"
        ],
        "summary": "Analyze",
        "operationId": "analyze_api_ai_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/explain": {
      "post": {
        "tags": [
          "AI integration"
        ],
        "summary": "Explain",
        "operationId": "explain_api_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/report/executive-brief": {
      "post": {
        "tags": [
          "AI integration"
        ],
        "summary": "Report",
        "operationId": "report_api_report_executive_brief_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Analysis": {
        "properties": {
          "source": {
            "type": "string",
            "const": "ai",
            "title": "Source",
            "default": "ai"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risks"
          },
          "tradeoffs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tradeoffs"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "strengths",
          "risks",
          "tradeoffs",
          "recommendations"
        ],
        "title": "Analysis"
      },
      "AnalyzeRequest": {
        "properties": {
          "simulation": {
            "$ref": "#/components/schemas/SimulationInput"
          },
          "question": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "simulation"
        ],
        "title": "AnalyzeRequest"
      },
...
      "Decision": {
        "properties": {
          "measureId": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Measureid"
          },
          "districtId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Districtid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "measureId"
        ],
        "title": "Decision"
      },
...
      "ErrorResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/Issue"
            },
            "type": "array",
            "title": "Errors"
          },
          "score": {
            "type": "null",
            "title": "Score"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message",
          "code"
        ],
        "title": "ErrorResponse"
      },
      "ExplainRequest": {
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Decisions"
          },
          "question": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decisions"
        ],
        "title": "ExplainRequest"
      },
...
      "Issue": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "measureIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Measureids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Issue"
      },
...
      "SimulationInput": {
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Decisions"
          }
        },
        "type": "object",
        "required": [
          "decisions"
        ],
        "title": "SimulationInput"
      },
...
    }
  }
}